Refactor 699.
